extract constant form information in CreateProfile to "form" file
